+++ EquineElegance.Entities/Feeder.cs
@@ -12,7 +12,7 @@ namespace EquineElegance.Entities
 
         public string Capacity { get; set; }
 
-        public Feeder(string name, string description, double price, string image, int amountInStock, string dimensions, string capacity) : 
+        public Feeder(string name, string description, decimal price, string image, int amountInStock, string dimensions, string capacity) : 
             base(name, description, price, image, amountInStock, dimensions)
         {
             Capacity = capacity;
