+++ EquineElegance.Entities/RiderProduct.cs
@@ -12,7 +12,7 @@ namespace EquineElegance.Entities
 
         public Color Color { get; set; }
 
-        public RiderProduct(string name, string description, double price, string image, int amountInStock, Color color)
+        public RiderProduct(string name, string description, decimal price, string image, int amountInStock, Color color)
             : base(name, description, price, image, amountInStock)
         {
             Color = color;
