+++ EquineElegance.Entities/RidingPants.cs
@@ -13,7 +13,7 @@ namespace EquineElegance.Entities
         public Gender Gender { get; set; }
         public PantsSize PantsSize { get; set; }
 
-        public RidingPants(string name, string description, double price, string image, int amountInStock, Color color, Gender gender, PantsSize pantsSize) : 
+        public RidingPants(string name, string description, decimal price, string image, int amountInStock, Color color, Gender gender, PantsSize pantsSize) : 
             base(name ,description, price, image, amountInStock, color)
         {
             Gender = gender;
