+++ EquineElegance.Entities/SaddlePad.cs
@@ -12,7 +12,7 @@ namespace EquineElegance.Entities
 
         public SaddlePadType SaddlePadType { get; set; }
 
-        public SaddlePad(string name, string description, double price, string image, int amountInStock, Color color, HorseSize horseSize, 
+        public SaddlePad(string name, string description, decimal price, string image, int amountInStock, Color color, HorseSize horseSize, 
             SaddlePadType saddlePadType) : base(name, description, price, image, amountInStock, color, horseSize)
         {
             SaddlePadType = saddlePadType;
